refactor(mutation): split endpoint resolution out of maintain.rs
The parallel-edge fix pushed `maintain.rs` to 2503 lines, three over the
source-quality ceiling. Rather than raise the ceiling, move the cluster
that resolves an `add_connections` frame's endpoint ids to node indices
— `ResolvedEndpoints`, `resolve_endpoints`, `scan_endpoint_rows`,
`resolve_pairs`, `title_column_indices`, `report_null_id_skips` — into
`mutation/endpoints.rs`. It is one job with one caller and no other user
in the workspace; `maintain.rs` drops to 2316 lines.

The initial-load decision moves onto `InitialLoad::owns_every_edge`, so
the renamed `add_connections_with_initial_load` measures one line and
two branches *below* the ceilings the pre-fix `add_connections` had; the
function-exception baseline records the rename, not a loosening. The
`connect` allowance gains the reason comment the allowance gate wants.
