+++ tests/api-baselines/source-quality.json
@@ -352,8 +352,8 @@
     },
     {
       "path": "crates/kglite/src/graph/mutation/maintain.rs",
-      "qualified_name": "crate::add_connections",
-      "lines": 250,
+      "qualified_name": "crate::add_connections_with_initial_load",
+      "lines": 249,
       "branches": 27,
       "nesting": 3
     },
